moves file viewer under UI folder
