+++ toolkit/popup/src/main/java/com/arcgismaps/toolkit/popup/internal/element/attachment/AttachmentTile.kt
@@ -60,8 +60,8 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.arcgismaps.LoadStatus
 import com.arcgismaps.mapping.popup.PopupAttachmentType
-import com.arcgismaps.toolkit.popup.internal.fileviewer.ViewableFile
-import com.arcgismaps.toolkit.popup.internal.fileviewer.toViewableFileType
+import com.arcgismaps.toolkit.popup.internal.ui.fileviewer.ViewableFile
+import com.arcgismaps.toolkit.popup.internal.ui.fileviewer.toViewableFileType
 import kotlinx.coroutines.flow.MutableStateFlow
 
 @Composable
